python/requirements.txt
python-dotenv>=0.19.0
requests>=2.31.0s`